Update(3.0.0): Fixed testing db handling in tests

+++ tests/main.py
@@ -3,7 +3,13 @@ import os
 from app.settings import config
 
 
+def _create_test_db_if_doesnt_exist():
+    if not os.path.exists(config.test_db_path):
+        os.mknod(config.test_db_path)
+
+
 def setup_testing():
+    _create_test_db_if_doesnt_exist()
     config.is_testing = True
     from app.db.accounts import users
     from app.db.categories import categories
@@ -13,4 +19,5 @@ def setup_testing():
 
 
 def clean_testing():
-    os.remove(config.test_db_path)
+    if os.path.exists(config.test_db_path):
+        os.remove(config.test_db_path)
